+++ RosalilaShooter/Stage/Layer.cpp
@@ -18,7 +18,7 @@ Layer::Layer(vector<LayerFrame*> layer_frames,int frame_duration,int depth_effec
 Layer::~Layer()
 {
     rosalila()->utility->writeLogLine("Deleting layer.");
-    for(int i=0;i<layer_frames.size();i++)
+    for(int i=0;i<(int)layer_frames.size();i++)
     {
         delete layer_frames[i];
     }
